Rename key_store to cursor_store
