+++ streams/src/api/cursor_store.rs
